refactor: add js highlight to the homework layout using hoc; move task-01 folder to the tasks folder
