+++ src/components/index.ts
@@ -1,2 +1,2 @@
 export * from './card';
-export * from './task-01';
+export * from './tasks';
