+++ src/components/tasks/task-01/index.ts
@@ -1,0 +1,1 @@
+export * from './task-01';
